Conflicts:
	crawl-ref/source/dat/descript/ability.txt
	crawl-ref/source/dat/descript/gods.txt
	crawl-ref/source/religion.cc

+++ crawl-ref/source/rltiles/tiledef_defines.h
@@ -3,8 +3,9 @@
 
 #include <cassert>
 #include <vector>
+#include <stdint.h>
 
-typedef unsigned int tileidx_t;
+typedef uint64_t tileidx_t;
 
 class tile_info
 {
